requirements.txt
# Core Data Science
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1

# Web Framework & Visualization
streamlit==1.28.0
plotly==5.18.0
streamlit-plotly-events==0.0.6

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.0
tensorflow==2.13.0
joblib==1.3.2

# Data Synthesis & Utilities
faker==20.0.0
python-dateutil==2.8.2

# APIs & Backend (for future use)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0

# Database (optional, for future)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Testing & Code Quality
pytest==7.4.2
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Utilities
PyYAML==6.0.1
tqdm==4.66.1

# Development Tools
ipython==8.17.2
jupyter==1.0.0
